+++ CodeStarterCamp_Week1/main.swift
@@ -8,24 +8,8 @@
 
 import Foundation
 
-let icecreamWidth = 11
+//makeIcecreamBody(ingredient: "*", height: 8)
+//makeStick(height: 4)
 
-func makeIcecreamBody(ingredient: String, height: Int) -> Void {
-    for _ in 1...height {
-        for _ in 1...icecreamWidth {
-            print("\(ingredient)", terminator: "")
-        }
-        print("")
-    }
-}
-func makeStick(height: Int) -> Void {
-    for _ in 1...height {
-        for _ in 1...icecreamWidth/2-1 {
-            print(" ", terminator: "")
-        }
-        print("| |")
-    }
-}
-
-makeIcecreamBody(ingredient: "*", height: 8)
-makeStick(height: 4)
+makePeperoStick(stickLength: 4)
+makePeperoBody(bodyLength: 4, function: setPeperoBody(body:topping:), peperoBody: "00", peperoTopping: "**")
